LayoutTests/fast/page-color-sampling/blend-modal-background-color-2-expected.txt

PASS colors.top is "rgb(153, 153, 153)"
PASS successfullyParsed is true

TEST COMPLETE
Hello
